Split the DummyModContainer from the IFMLLoadingPlugin to avoid unecessary classloading of certain classes. Fixes #2482.
